Start upon boot for data source device
We were previously using bluetooth state changes as events to trigger the activation of
bluetooth data sharing service. This approach failed to work upon boot, since it boots
up with the bluetooth in the same state as it powered off in.

Now we will watch for a "boot completed" and use that also to activate tethering service.