China broke WifiManager.setWifiEnabled(boolean)
Setting the parameter "false" is supposed to DISABLE wifi, not toggle it.
So what was happening is when bluetooth connected, it would toggle wifi. Connect bluetooth --> wifi enables. Connect bluetooth --> wifi disables. Connect bluetooth --> wifi enables. Not good.

Instead, what we do here is set up a broadcast receiver for wifi state changes, and if the new state is "enabled", then we run a disable.

It is no longer linked to PAN, but basically makes it impossible to enable wifi. It is necessary to remember to disable this "forced disable" mode if you need to temporarily re-enable wifi.